MNT move _sage_tags in _tags module
