+++ sklearn/feature_selection/_rfe.py
@@ -10,9 +10,10 @@ import numpy as np
 import numbers
 from joblib import Parallel, effective_n_jobs
 
-from ..utils import _safe_tags
+
 from ..utils.metaestimators import if_delegate_has_method
 from ..utils.metaestimators import _safe_split
+from ..utils._tags import _safe_tags
 from ..utils.validation import check_is_fitted
 from ..utils.validation import _deprecate_positional_args
 from ..utils.fixes import delayed
