+++ sklearn/model_selection/_search.py
@@ -33,11 +33,9 @@ from ._validation import _insert_error_scores
 from ._validation import _normalize_score_results
 from ..exceptions import NotFittedError
 from joblib import Parallel
-from ..utils import (
-    check_random_state,
-    _safe_tags,
-)
+from ..utils import check_random_state
 from ..utils.random import sample_without_replacement
+from ..utils._tags import _safe_tags
 from ..utils.validation import indexable, check_is_fitted, _check_fit_params
 from ..utils.validation import _deprecate_positional_args
 from ..utils.metaestimators import if_delegate_has_method
