+++ sklearn/utils/estimator_checks.py
@@ -51,10 +51,8 @@ from ..model_selection import ShuffleSplit
 from ..model_selection._validation import _safe_split
 from ..metrics.pairwise import (rbf_kernel, linear_kernel, pairwise_distances)
 
-from .import (
-    shuffle,
-    _safe_tags,
-)
+from .import shuffle
+from ._tags import _safe_tags
 from .validation import has_fit_parameter, _num_samples
 from ..preprocessing import StandardScaler
 from ..preprocessing import scale
