+++ sklearn/utils/tests/test_utils.py
@@ -8,14 +8,12 @@ import pytest
 import numpy as np
 import scipy.sparse as sp
 
-from sklearn.base import BaseEstimator
 from sklearn.utils._testing import (assert_array_equal,
                                     assert_allclose_dense_sparse,
                                     assert_warns_message,
                                     assert_no_warnings,
                                     _convert_container)
 from sklearn.utils import check_random_state
-from sklearn.utils import _DEFAULT_TAGS
 from sklearn.utils import _determine_key_type
 from sklearn.utils import deprecated
 from sklearn.utils import gen_batches
@@ -24,7 +22,6 @@ from sklearn.utils import resample
 from sklearn.utils import safe_mask
 from sklearn.utils import column_or_1d
 from sklearn.utils import _safe_indexing
-from sklearn.utils import _safe_tags
 from sklearn.utils import shuffle
 from sklearn.utils import gen_even_slices
 from sklearn.utils import _message_with_time, _print_elapsed_time
@@ -696,63 +693,3 @@ def test_to_object_array(sequence):
     assert isinstance(out, np.ndarray)
     assert out.dtype.kind == 'O'
     assert out.ndim == 1
-
-
-class EstimatorNoTags:
-    pass
-
-
-class EstimatorOwnTags:
-    def _get_tags(self):
-        return {}
-
-
-@pytest.mark.parametrize(
-    "estimator, err_msg",
-    [
-        (BaseEstimator(), "The key xxx is neither defined"),
-        (EstimatorNoTags(), "The key xxx is not a default tags defined"),
-    ],
-)
-def test_safe_tags_error(estimator, err_msg):
-    # Check that safe_tags raises error in ambiguous case.
-    with pytest.raises(ValueError, match=err_msg):
-        _safe_tags(estimator, key="xxx")
-
-
-@pytest.mark.parametrize(
-    "estimator, key, default, expected_results",
-    [
-        (EstimatorNoTags(), None, None, _DEFAULT_TAGS),
-        (EstimatorNoTags(), "allow_nan", None, _DEFAULT_TAGS["allow_nan"]),
-        (
-            EstimatorNoTags(),
-            "allow_nan",
-            not _DEFAULT_TAGS["allow_nan"],
-            not _DEFAULT_TAGS["allow_nan"],
-        ),
-        (EstimatorNoTags(), "xxx", True, True),
-        (BaseEstimator(), None, None, _DEFAULT_TAGS),
-        (BaseEstimator(), "allow_nan", None, _DEFAULT_TAGS["allow_nan"]),
-        (
-            BaseEstimator(),
-            "allow_nan",
-            not _DEFAULT_TAGS["allow_nan"],
-            _DEFAULT_TAGS["allow_nan"],
-        ),
-        (BaseEstimator(), "xxx", True, True),
-        (EstimatorOwnTags(), None, None, _DEFAULT_TAGS),
-        (EstimatorOwnTags(), "allow_nan", None, _DEFAULT_TAGS["allow_nan"]),
-        (
-            EstimatorOwnTags(),
-            "allow_nan",
-            not _DEFAULT_TAGS["allow_nan"],
-            not _DEFAULT_TAGS["allow_nan"],
-        ),
-        (EstimatorOwnTags(), "xxx", True, True),
-    ],
-)
-def test_safe_tags_no_get_tags(estimator, key, default, expected_results):
-    # check the behaviour of _safe_tags when an estimator does not implement
-    # _get_tags
-    assert _safe_tags(estimator, key=key, default=default) == expected_results
